move the get_monitoring_config bits to a separate script
i want this to be separate from the main
create_tables_and_add_defaults stuff. It's cleaner this way.

Also flatten the _update_config function one level right there.
