+++ cfy_manager/components/restservice/restservice.py
@@ -268,12 +268,9 @@ class RestService(BaseComponent):
         if config[MANAGER][HOSTNAME] in managers:
             db.update_stored_manager(configs)
         else:
-            cluster_cfg_fn, rabbitmq_ca_fn = db.insert_manager(configs)
+            db.insert_manager(configs)
             if len(managers) > 0:
                 self._join_cluster(configs)
-                if MONITORING_SERVICE in config.get(SERVICES_TO_INSTALL):
-                    self._prepare_cluster_config_update(cluster_cfg_fn,
-                                                        rabbitmq_ca_fn)
 
     def _initialize_db(self, configs):
         logger.info('DB not initialized, creating DB...')
