+++ cfy_manager/components/restservice/scripts/create_tables_and_add_defaults.py
@@ -17,7 +17,6 @@
 from __future__ import print_function
 
 import os
-from os.path import isfile
 import sys
 import json
 import atexit
@@ -168,35 +167,6 @@ def _insert_manager(config):
     sm.put(inst)
 
 
-def _prepare_config_for_monitoring():
-    sm = get_storage_manager()
-    cfg = {}
-    rabbitmq_nodes = sm.list(models.RabbitMQBroker)
-    if len(rabbitmq_nodes) > 0:
-        tmp_ca_cert_path = rabbitmq_nodes[0].write_ca_cert()
-        if isfile(tmp_ca_cert_path):
-            RETURN_DICT['rabbitmq_ca_cert_path'] = tmp_ca_cert_path
-        cfg['rabbitmq'] = {
-            'cluster_members': {}
-        }
-        for node in rabbitmq_nodes:
-            cfg['rabbitmq']['cluster_members'][node.name] = {
-                'networks': {'default': node.private_ip}
-            }
-    postgresql_server_nodes = sm.list(models.DBNodes)
-    if len(postgresql_server_nodes) > 0:
-        cfg['postgresql_server'] = {'cluster': {'nodes': {}}}
-        for node in postgresql_server_nodes:
-            cfg['postgresql_server']['cluster']['nodes'][node.name] = {
-                'ip': node.private_ip
-            }
-    if not cfg:
-        return
-    with tempfile.NamedTemporaryFile(delete=False, mode='w') as fp:
-        json.dump(cfg, fp)
-        RETURN_DICT['cluster_nodes_config'] = fp.name
-
-
 def _insert_cert(cert, name):
     sm = get_storage_manager()
     inst = models.Certificate(
@@ -315,7 +285,5 @@ if __name__ == '__main__':
         _insert_db_nodes(script_config['db_nodes'])
     if script_config.get('usage_collector'):
         _insert_usage_collector(script_config['usage_collector'])
-    if script_config.get('manager'):
-        _prepare_config_for_monitoring()
 
     print(json.dumps(RETURN_DICT))
